things are broken help
