+++ Underloading/ArrayUtils/SortAscending.cs
@@ -21,7 +21,7 @@ namespace ArrayUtils
                     }
                 }
             }
-            string xString = ArrayToString(x);
+            string xString = thing.ArrayToString(x);
             return xString;
         }
 
@@ -41,48 +41,8 @@ namespace ArrayUtils
                     }
                 }
             }
-            string xString = ArrayToString(x);
+            string xString = thing.ArrayToString(x);
             return xString;
-        }
-
-        public static string ArrayToString(int[] x)
-        {
-            string arrayToString = "{ ";
-            int i = 0;
-
-            foreach (int item in x)
-            {
-                if (i < x.Length - 1)
-                {
-                    arrayToString += $"{Convert.ToString(x.GetValue(i))}, ";
-                }
-                else
-                {
-                    arrayToString += $"{Convert.ToString(x.GetValue(i))} }}";
-                }
-                i++;
-            }
-            return arrayToString;
-        }
-
-        public static string ArrayToString(double[] x)
-        {
-            string arrayToString = "{ ";
-            int i = 0;
-
-            foreach (int item in x)
-            {
-                if (i < x.Length - 1)
-                {
-                    arrayToString += $"{Convert.ToString(x.GetValue(i))}, ";
-                }
-                else
-                {
-                    arrayToString += $"{Convert.ToString(x.GetValue(i))} }}";
-                }
-                i++;
-            }
-            return arrayToString;
         }
     }
 }
