+++ Underloading/Program.cs
@@ -10,40 +10,43 @@ namespace Underloading
             int[] intArr = {1, 9, 4, 18, 7, 53};
             double[] doubleArr = {1.1, 9.9, 4.4, 18.18, 7.7, 53.53};
 
-            double intAvg = averaging.Average(intArr);
-            Console.WriteLine("The average of your int array is: " + intAvg);
-            double doubleAvg = averaging.Average(doubleArr);
-            Console.WriteLine("The average of your double array is: " + doubleAvg);
+            string fard = duplicator.Dupe(intArr);
+            Console.WriteLine(fard);
 
-            int intMax = macks.Max(intArr);
-            Console.WriteLine("The max of your int array is: " + intMax);
-            double doubleMax = macks.Max(doubleArr);
-            Console.WriteLine("The max of your double array is: " + doubleMax);
+            // double intAvg = averaging.Average(intArr);
+            // Console.WriteLine("The average of your int array is: " + intAvg);
+            // double doubleAvg = averaging.Average(doubleArr);
+            // Console.WriteLine("The average of your double array is: " + doubleAvg);
 
-            int intMin = minnie.Min(intArr);
-            Console.WriteLine("The min of your int array is: " + intMin);
-            double doubleMin = minnie.Min(doubleArr);
-            Console.WriteLine("The min of your double array is: " + doubleMin);
+            // int intMax = macks.Max(intArr);
+            // Console.WriteLine("The max of your int array is: " + intMax);
+            // double doubleMax = macks.Max(doubleArr);
+            // Console.WriteLine("The max of your double array is: " + doubleMax);
 
-            int intProduct = productivity.Product(intArr);
-            Console.WriteLine("The product of your int array is: " + intProduct);
-            int doubleProduct = (int)productivity.Product(doubleArr);
-            Console.WriteLine("The product of your double array is: " + doubleProduct);
+            // int intMin = minnie.Min(intArr);
+            // Console.WriteLine("The min of your int array is: " + intMin);
+            // double doubleMin = minnie.Min(doubleArr);
+            // Console.WriteLine("The min of your double array is: " + doubleMin);
 
-            int intSum = summarizing.Sum(intArr);
-            Console.WriteLine("The sum of your int array is: " + intSum);
-            int doubleSum = (int)summarizing.Sum(doubleArr);
-            Console.WriteLine("The sum of your double array is: " + doubleSum);
+            // int intProduct = productivity.Product(intArr);
+            // Console.WriteLine("The product of your int array is: " + intProduct);
+            // int doubleProduct = (int)productivity.Product(doubleArr);
+            // Console.WriteLine("The product of your double array is: " + doubleProduct);
 
-            string intSortUp = ascent.SortAscending(intArr);
-            Console.WriteLine("Your int array sorted going up is: " + intSortUp);
-            string doubleSortUp = ascent.SortAscending(doubleArr);
-            Console.WriteLine("Your double array sorted going up is: " + doubleSortUp);
+            // int intSum = summarizing.Sum(intArr);
+            // Console.WriteLine("The sum of your int array is: " + intSum);
+            // int doubleSum = (int)summarizing.Sum(doubleArr);
+            // Console.WriteLine("The sum of your double array is: " + doubleSum);
 
-            string intSortDown = descent.SortDescending(intArr);
-            Console.WriteLine("Your int array sorted going downwards is: " + intSortDown);
-            string doubleSortDown = descent.SortDescending(doubleArr);
-            Console.WriteLine("Your double array sorted going downwards is: " + doubleSortDown);
+            // string intSortUp = ascent.SortAscending(intArr);
+            // Console.WriteLine("Your int array sorted going up is: " + intSortUp);
+            // string doubleSortUp = ascent.SortAscending(doubleArr);
+            // Console.WriteLine("Your double array sorted going up is: " + doubleSortUp);
+
+            // string intSortDown = descent.SortDescending(intArr);
+            // Console.WriteLine("Your int array sorted going downwards is: " + intSortDown);
+            // string doubleSortDown = descent.SortDescending(doubleArr);
+            // Console.WriteLine("Your double array sorted going downwards is: " + doubleSortDown);
         }
     }
 }
